Implement tab activity detection to optimize  API calls (#112)
### **Overview:**
This resolves #105 
Previously, the application was fetching the price at regular intervals,
regardless of whether the user's tab was active or not. This approach
led to reaching the daily limit of the free Infura API when deployed for
Sepolia, as a single user could leave a tab open, consuming the limit.

Implemented a solution that detects whether the tab is active or not.
The price update logic now only executes when the tab is active,
significantly reducing the number of API calls and preventing the daily
limit from being reached.

This change ensures that the application remains functional and
efficient under the constraints of the free Infura API, while also
improving the user experience by reducing unnecessary data fetching when
the application is not in use.

###  Detailed Description:

| Before Changes | After Changes |
|:---:|:---:|
| ![before
changes](https://github.com/DjedAlliance/Djed-Solidity-WebDashboard/assets/157852043/22baf38c-a0d7-4477-9169-c0abd05a22b1)
| ![after
changes](https://github.com/DjedAlliance/Djed-Solidity-WebDashboard/assets/157852043/317f029a-93c7-4f09-8f1b-5e5e224a1fe6)
|

As shown in the images ,Previously, even when the tab was not active and
the user had switched to another tab, there were **10 request methods**
being called every minute. This behaviour could easily lead to exceeding
the API rate limit.

implementing a solution that involves detecting whether the tab is
active. As a result, **no requests (zero)** are being made when the tab
is not active.