Revert the changes made in the sepolia-frontend (#116)
Reverting the changes made to display price feeds for the Sepolia
deployment in the etherum.js file.
#100 also led to increase in code duplication within the etherum.js
file.We will have to find a solution to display the prices efficiently
for different networks.

Update: Deployed the djed contract with modified parameters and updated
the addresses.

In `ethereum.js` ,Using === to compare a BigNumber with a primitive
number will always return false, which can cause logical errors in our
code. Therefore, we use the futureScPrice.isZero() method provided by
the ethers.js library to handle BigNumber comparisons correctly.